Moved text splitters inside of the components folder
